refactor(tv): consume :core-network DTOs instead of forking
Mirrors what RFC 0004 did for :wear: :tv now depends on :core-network and
reads/writes the canonical EventDto + WebcastDto, dropping its local trimmed
forks in tv/data/api/TbaDto.kt. Behavior is unchanged — the tv version was a
strict subset of fields and the WebcastDto shapes were identical.

- Add implementation(project(":core-network")) to :tv.
- Move the toDomainOrNull mapper into its own EventMapper.kt; the .orEmpty()
  reflects that canonical EventDto.webcasts is List<WebcastDto>? (was a
  non-null default-empty list on the tv fork).
- Backfill event_code + year on every entry in events_fixture.json so the
  bundled-asset path satisfies canonical EventDto's required fields. Derived
  mechanically from the existing key, e.g. "2026mibatb" → event_code "mibatb",
  year 2026 — same values the live API returns.
- Lower :core-network minSdk from 26 to 23 to match :tv's floor. The module
  is pure kotlinx-serialization data classes; no version-dependent APIs.
- Replace the four EventDto(...) call sites in EventLogicTest with a small
  dto(...) helper that hides the new required fields, keeping each test
  focused on the field it exercises.

Verified locally: :app/:wear/:tv all assemble debug+release, :tv unit tests
+ lint green, :tv smoke-tested in the emulator — events render identically
through the canonical DTOs (live API and bundled fixture paths both clean).
