+++ tv/build.gradle.kts
@@ -163,6 +163,9 @@ play {
 }
 
 dependencies {
+    // Shared modules
+    implementation(project(":core-network"))
+
     // Compose
     implementation(platform(libs.compose.bom))
     implementation(libs.compose.ui)
